Renamed homepage to itesco page
