+++ app.py
@@ -17,8 +17,9 @@ app.config.suppress_callback_exceptions = True
 # why are you here my friend?
 
 from apps import homepage
+from apps import itesco_page
 from db_connection import conn
-from callbacks.callbacks_homepage import rc_itesco_weighted_evo_graph, rc_itesco_main_cat_graph2, rc_itesco_drilldown_sub, rc_itesco_main_cat_graph_agg, rc_itesco_product_graph_dropdown
+from callbacks.callbacks_itesco_page import rc_itesco_weighted_evo_graph, rc_itesco_main_cat_graph2, rc_itesco_drilldown_sub, rc_itesco_main_cat_graph_agg, rc_itesco_product_graph_dropdown
 
 app.layout = html.Div([
     dcc.Location(id='url', refresh=False),
@@ -30,6 +31,8 @@ app.layout = html.Div([
 def display_page(pathname):
     if pathname == '/':
         return homepage.layout
+    if pathname == '/itesco':
+        return itesco_page.layout
     # elif pathname == '/kos':
     #     return kos.layout # --> add!
     # elif pathname == '/zajimavosti':
